[#939] Disabled else-after-return check
This check has unintended effects when combined with MISRA C++:2023 Rule
9.4.1, which mandates a (possibly empty) trailing else after every
if..else if. In combination, these two rules effectively prohibit the
use of control statements like return/continue/break in an else if
block, as it is not possible to comply with both rules at the same time
in such a situation.